requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
python-multipart>=0.0.20,<1.0
PyPDF2>=3.0,<4.0
python-dotenv>=1.0,<2.0
pydantic-settings>=2.0,<3.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
openai>=2.33.0,<3.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
